+++ src/schema/schema-generator.ts
@@ -686,19 +686,29 @@ export abstract class SchemaGenerator {
   ): GraphQLFieldConfigArgumentMap {
     return params!.reduce<GraphQLFieldConfigArgumentMap>((args, param) => {
       if (param.kind === "arg") {
+        const type = this.getGraphQLInputType(
+          target,
+          propertyName,
+          param.getType(),
+          param.typeOptions,
+          param.index,
+          param.name,
+        );
+        const argDirectives = getMetadataStorage()
+          .argumentDirectives.filter(
+            it =>
+              it.target === target &&
+              it.fieldName === propertyName &&
+              it.parameterIndex === param.index,
+          )
+          .map(it => it.directive);
         // eslint-disable-next-line no-param-reassign
         args[param.name] = {
           description: param.description,
-          type: this.getGraphQLInputType(
-            target,
-            propertyName,
-            param.getType(),
-            param.typeOptions,
-            param.index,
-            param.name,
-          ),
+          type,
           defaultValue: param.typeOptions.defaultValue,
           deprecationReason: param.deprecationReason,
+          astNode: getInputValueDefinitionNode(param.name, type, argDirectives),
         };
       } else if (param.kind === "args") {
         const argumentType = getMetadataStorage().argumentTypes.find(