+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - **Breaking Change**: expose shim as a package entry point `type-graphql/shim` (and `/node_modules/type-graphql/build/typings/shim.ts`)
 - support defining directives on `@Field` of `@Args`
+- support defining directives on inline `@Arg`
 
 ## v2.0.0-beta.3